+++ CHANGELOG.md
@@ -18,6 +18,7 @@ react-testing-library and wrap the test component in most of
 the common application contexts (grommet, intl, react-router, 
 react-query, react-hot-toast, breadcrumb, styled-components)
 - Add a link to generate an ics file on the scheduled page
+- Manage tags on video resource
 
 ### Changed